fix validateRole
