+++ backend/src/flox/modules/roles/config.ts
@@ -1,6 +1,8 @@
-import { mergeConfigurations } from '../../core/flox-helpers';
+import {
+  floxModuleOptions,
+  mergeConfigurations,
+} from '../../core/flox-helpers';
 import { MODULES } from '../../MODULES';
-import { floxModuleOptions } from '../../flox';
 
 /**
  * The roles module includes role management functionalities and decorators for restricting certain resources to certain
